+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+* Add isolate support with `isolate` extension method
+* Add `map` and `onIdle` extension methods
+
 ## 1.0.1
 
 * Update SDK constraints in pubspec.yaml